# apps/api/sdks/typescript/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Janua API",
    "description": "Modern authentication and identity platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/mfa/status": {
      "get": {
        "tags": [
          "mfa"
        ],
        "summary": "Get Mfa Status",
        "description": "Get MFA status for current user",
        "operationId": "get_mfa_status_api_v1_mfa_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MFAStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mfa/enable": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Enable Mfa",
        "description": "Enable MFA for current user",
        "operationId": "enable_mfa_api_v1_mfa_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAEnableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MFAEnableResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mfa/verify": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Verify Mfa",
        "description": "Verify MFA setup with TOTP code",
        "operationId": "verify_mfa_api_v1_mfa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mfa/disable": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Disable Mfa",
        "description": "Disable MFA for current user",
        "operationId": "disable_mfa_api_v1_mfa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFADisableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/mfa/regenerate-backup-codes": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Regenerate Backup Codes",
        "description": "Regenerate backup codes",
        "operationId": "regenerate_backup_codes_api_v1_mfa_regenerate_backup_codes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MFABackupCodesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mfa/validate-code": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Validate Mfa Code",
        "description": "Validate an MFA code (for testing/verification)",
        "operationId": "validate_mfa_code_api_v1_mfa_validate_code_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mfa/recovery-options": {
      "get": {
        "tags": [
          "mfa"
        ],
        "summary": "Get Recovery Options",
        "description": "Get MFA recovery options for a user (public endpoint)",
        "operationId": "get_recovery_options_api_v1_mfa_recovery_options_get",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mfa/initiate-recovery": {
      "post": {
        "tags": [
          "mfa"
        ],
        "summary": "Initiate Mfa Recovery",
        "description": "Initiate MFA recovery process",
        "operationId": "initiate_mfa_recovery_api_v1_mfa_initiate_recovery_post",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MFABackupCodesResponse": {
        "properties": {
          "backup_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Backup Codes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "backup_codes",
          "generated_at"
        ],
        "title": "MFABackupCodesResponse",
        "description": "MFA backup codes response"
      },
      "MFADisableRequest": {
        "properties": {
          "password": {
            "type": "string",
            "title": "Password"
          },
          "code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 6,
                "minLength": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "MFADisableRequest",
        "description": "MFA disable request"
      },
      "MFAEnableRequest": {
        "properties": {
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "MFAEnableRequest",
        "description": "MFA enable request"
      },
      "MFAEnableResponse": {
        "properties": {
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "qr_code": {
            "type": "string",
            "title": "Qr Code"
          },
          "backup_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Backup Codes"
          },
          "provisioning_uri": {
            "type": "string",
            "title": "Provisioning Uri"
          }
        },
        "type": "object",
        "required": [
          "secret",
          "qr_code",
          "backup_codes",
          "provisioning_uri"
        ],
        "title": "MFAEnableResponse",
        "description": "MFA enable response"
      },
      "MFAStatusResponse": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "backup_codes_remaining": {
            "type": "integer",
            "title": "Backup Codes Remaining"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "verified",
          "backup_codes_remaining",
          "last_used_at"
        ],
        "title": "MFAStatusResponse",
        "description": "MFA status response"
      },
      "MFAVerifyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "MFAVerifyRequest",
        "description": "MFA verification request"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}